Prevent installation of tests through setup.py
Currently when the package is built and installed into the system
using the python3 setup.py/python3 setup.py install flow, it also
installs the tests. There's no real reason the tests should be
distributed with the rest of the code. This patch prevents the tests
from being installed in this manner.